added dir for helpfunctions
