+++ controllers/bike.js
@@ -27,30 +27,7 @@ const createBike = async (req, res) => {
     }
 };
 
-async function addbikes(longmin, longmax, latmin, latmax) {
-    let named = 'Bike-'
 
-    for (var i = 1; i < 100; i++) {
-        let lat = Math.random() * (latmax - latmin) + latmin;
-        let long = Math.random() * (longmax - longmin) + longmin;
-        let name = named + i.toString()
-        await Bike.create({
-            name, maxspeed: 30, speed: 0, batterylevel: 100, history: [], location: {
-                type: "Point",
-                coordinates: [
-                    long,
-                    lat
-
-                ]
-            }, inCity: "6378989b6a6403d2a9c6edb1"
-        }
-        );
-    }
-};
-
-async function deleteBikes() {
-    await Bike.deleteMany({ active: false })
-}
 
 
 
@@ -59,7 +36,5 @@ async function deleteBikes() {
 module.exports = {
     createBike,
     getAllBikes,
-    addbikes,
-    getOneBike,
-    deleteBikes
+    getOneBike
 }
