+++ models/User.js
@@ -1,0 +1,2 @@
+const mongoose = require('mongoose');
+
